pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "owlai"
version = "0.3.3"
description = "OwlAI: AI agent with RAG capabilities"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "OwlAI Team"}
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
    "langchain-core>=0.3,<0.4",
    "langchain-text-splitters>=0.3,<0.4",
    "langchain-community>=0.3,<0.4",
    "langchain-openai>=0.0.5",
    "langchain-anthropic>=0.1.1",
    "langgraph>=0.0.22",
    "langchain-huggingface>=0.0.2",
    "pydantic>=2.5.2",
    "openai",
    "anthropic",
    "numpy",
    "torch",
    "torchaudio",
    "rich",
    "python-dotenv",
    "PyYAML",
    "transformers",
    "sentence-transformers",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "psutil>=5.9.0",
    "GPUtil>=1.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "mypy",
]
tts = [
    "coqui-tts",
    "pyttsx3",
    "edge-tts",
    "elevenlabs",
    "gTTS",
]
spotify = [
    "spotipy",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["owlai*"]
namespaces = false

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
filterwarnings = [
    "ignore::DeprecationWarning:pythonjsonlogger.*:",
    "ignore::DeprecationWarning:importlib._bootstrap:",
    "ignore::DeprecationWarning:numpy.core.*:",
    "ignore::DeprecationWarning:faiss.*:",
] 